requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
PyJWT==2.8.0
uuid==1.30
django-cors-headers==4.3.1
setuptools>=65.0.0